use config.go
