+++ formcentral/server.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-
-	"github.com/caarlos0/env/v6"
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/labstack/echo/v4"
 	"github.com/vlab-research/trans"
@@ -29,7 +27,6 @@ func (s *Server) GetTranslator(c echo.Context) error {
 	}
 
 	translator, err := trans.MakeTranslatorByShape(src, dest)
-
 	if err != nil {
 		msg := err.Error()
 		return echo.NewHTTPError(http.StatusInternalServerError, fmt.Sprintf("Could not create translation mapping. Failed with the following error: %v", msg))
@@ -46,12 +43,11 @@ type TranslatorRequest struct {
 
 func (s *Server) CreateTranslator(c echo.Context) error {
 	req := new(TranslatorRequest)
-	var dest *trans.FormJson
-
 	if err := c.Bind(req); err != nil {
 		return err
 	}
 
+	var dest *trans.FormJson
 	if req.Self {
 		dest = req.Form
 	} else {
@@ -72,21 +68,6 @@ func (s *Server) CreateTranslator(c echo.Context) error {
 	return c.JSON(http.StatusOK, translator)
 }
 
-type Config struct {
-	Db       string `env:"CHATBASE_DATABASE,required"`
-	User     string `env:"CHATBASE_USER,required"`
-	Password string `env:"CHATBASE_PASSWORD,required"`
-	Host     string `env:"CHATBASE_HOST,required"`
-	Port     string `env:"CHATBASE_PORT,required"`
-}
-
-func getConfig() Config {
-	cfg := Config{}
-	err := env.Parse(&cfg)
-	handle(err)
-	return cfg
-}
-
 func handle(err error) {
 	if err != nil {
 		log.Fatal(err)
@@ -94,9 +75,7 @@ func handle(err error) {
 }
 
 func main() {
-	cfg := getConfig()
-	pool := getPool(&cfg)
-
+	pool := getPool()
 	server := &Server{pool}
 
 	e := echo.New()
