+++ formcentral/test_helpers.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/jackc/pgconn"
@@ -15,16 +14,4 @@ func mustExec(t testing.TB, conn *pgxpool.Pool, sql string, arguments ...interfa
 		t.Fatalf("Exec unexpectedly failed with %v: %v", sql, err)
 	}
 	return
-}
-
-func testPool() *pgxpool.Pool {
-	url := fmt.Sprintf("postgres://%s:%s@%s:%s/%s", PostgresUser, PostgresPassword, PostgresHost, PostgresPort, PostgresDb)
-	config, err := pgxpool.ParseConfig(url)
-	handle(err)
-
-	ctx := context.Background()
-	pool, err := pgxpool.ConnectConfig(ctx, config)
-	handle(err)
-
-	return pool
 }
